Simplifies resolvers and updates dependencies
Many resolvers are not needed, or scoped globally, but only used for a
couple projects. In any case, the configuration was complex. This
starts over, adding the minimal repositories to get the build to pass.

This also updates certain dependencies, which in some cases makes them
available in standard repositories. These dependencies were checked to
ensure they work on JRE 7.